OK , getting there. ** I ** Can build the excel addin with VS 9 Pro and VS 8 Pro effortlessly. Most of the functions  in the xlw autogenerated addin have been implemented in C# and when the addin is built most of the Example.xls from the example in 'vanilla'  xlw will run. HOWEVER In VS8 ( and actualy in VS9 but to a much lesser extent) some fiddling is required with project settings, tweaking, adding the correct assemble depending on whether its Debug or Release etc. For ME this fiddling is effortless, BUT thats because I have done it 100's of times, but its not good enough for the end user.
Most of the BULK of this project is  done, that is we can write C# functions , press build , copy a couple of files and hey presto you have an excel addin (.xll) that provides those functions. 
What remains is really setting up a smooth workflow for the end user, buts its not trivial, as the project spans 2 runtimes, C++ & C#. Setting up a smooth workflow procees under the express editions is going to be even tricker as you cannot have VS Solutions having both C# AND C++ projects under the same umbrella. 

I think in the end the effort is going to be worth it. At work I have had C# excel addins built with xlwDotNet in production for about 5 month, on live Traders pricing sheets with Reuters feed computing C# function 30 times a minute the whole day and working very smoothly ;-)