OK, some progress. One of the issue was to link to the right xlwDotNet---.dll depending on whether in Debug or Release. Now have been able to do this but using the #using field in the project setting under C++ and it seems to be working nicely.
Next step is definitely the Template. A Director of project files that can just be copied to some place of the user's preference and the user can just just adding C# function and building an addin.  The usewr will however have to set the enviroment variable XLW to the directory where xlw is installed.  Is it possible to do that as part of the installation process of xlw ?