This is wierd.
The Build broke only for 64bit Release. Some obscure error:
error LNK2016: Absolute symbol '@comp.id' used as target of REL32 relocation

After googling found out that if you turn off 'Whole Program Optimzation' then it should work. Tried it .. and it worked. But still not explained. Something for JOhn to look at. Maybe we need to turn this option off on all VS2005 64bit Release builds just to be in the safe side.
Another curious thing, RTTI seems to be switched on by default on many builds .. should we turn it off. ?